# backend/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.ContadorEstados:
    properties:
      en_alerta:
        type: integer
      excedidos:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.EstadoPresupuesto:
    properties:
      categoria_id:
        type: string
      color:
        type: string
      disponible:
        type: number
      estado:
        type: string
      gastado:
        type: number
      monto_limite:
        type: number
      nombre:
        type: string
      porcentaje_usado:
        type: number
      presupuesto_id:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.GastoPorCategoria:
    properties:
      cantidad:
        type: integer
      categoria_id:
        type: string
      color:
        type: string
      icono:
        type: string
      nombre:
        type: string
      porcentaje:
        type: number
      total:
        type: number
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.Periodo:
    properties:
      anio:
        type: integer
      mes:
        type: integer
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.PuntoTendencia:
    properties:
      balance:
        type: number
      cantidad:
        type: integer
      etiqueta:
        type: string
      gastos:
        type: number
      ingresos:
        type: number
      periodo:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Periodo'
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.Resumen:
    properties:
      balance:
        type: number
      gastos:
        type: number
      ingresos:
        type: number
      movimientos:
        type: integer
      periodo:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Periodo'
      presupuestos:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.ContadorEstados'
      saldo_total:
        type: number
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.SaldoCuenta:
    properties:
      archivada:
        type: boolean
      color:
        type: string
      cuenta_id:
        type: string
      gastos:
        type: number
      ingresos:
        type: number
      nombre:
        type: string
      saldo:
        type: number
      saldo_inicial:
        type: number
      tipo:
        type: string
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle:
    properties:
      codigo:
        type: string
      detalles:
        items:
          type: string
        type: array
      mensaje:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta:
    properties:
      limite:
        type: integer
      pagina:
        type: integer
      total:
        type: integer
      total_paginas:
        type: integer
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre:
    properties:
      datos: {}
      meta:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta'
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError:
    properties:
      error:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API REST de finanzas personales: cuentas, categorias, transacciones, presupuestos y reportes.
    Todos los datos pertenecen a un usuario. El identificador del dueño sale siempre del token, nunca del cuerpo de la peticion.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinTrack API
  version: 0.1.0
paths:
...
  /reportes/estado-presupuestos:
    get:
      description: Presupuestado contra gastado en cada categoria del mes, con el
        semaforo ok / alerta / excedido. Cruza presupuestos con categorias y transacciones.
      parameters:
      - description: Mes del 1 al 12 (por defecto, el mes en curso)
        in: query
        name: mes
        type: integer
      - description: Año de 2000 a 2100 (por defecto, el año en curso)
        in: query
        name: anio
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.EstadoPresupuesto'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Estado de presupuestos (consulta relacional 2)
      tags:
      - reportes
  /reportes/gastos-por-categoria:
    get:
      description: 'En que se fue el dinero del mes: total, numero de movimientos
        y peso porcentual de cada categoria. Cruza transacciones con categorias.'
      parameters:
      - description: Mes del 1 al 12 (por defecto, el mes en curso)
        in: query
        name: mes
        type: integer
      - description: Año de 2000 a 2100 (por defecto, el año en curso)
        in: query
        name: anio
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.GastoPorCategoria'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Gastos por categoria (consulta relacional 1)
      tags:
      - reportes
  /reportes/resumen:
    get:
      description: Ingresos, gastos, balance, saldo disponible y conteo del semaforo
        de presupuestos.
      parameters:
      - description: Mes del 1 al 12 (por defecto, el mes en curso)
        in: query
        name: mes
        type: integer
      - description: Año de 2000 a 2100 (por defecto, el año en curso)
        in: query
        name: anio
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Resumen'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Resumen del mes
      tags:
      - reportes
  /reportes/saldos:
    get:
      description: 'saldo_inicial + ingresos - gastos. No se guarda: se calcula cruzando
        cuentas con transacciones.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.SaldoCuenta'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Saldo de cada cuenta
      tags:
      - reportes
  /reportes/tendencia:
    get:
      description: Serie de los ultimos meses terminando en el periodo pedido. Los
        meses sin movimientos salen con ceros.
      parameters:
      - description: Mes del 1 al 12 (por defecto, el mes en curso)
        in: query
        name: mes
        type: integer
      - description: Año de 2000 a 2100 (por defecto, el año en curso)
        in: query
        name: anio
        type: integer
      - description: Cuantos meses, de 1 a 24 (por defecto 6)
        in: query
        name: meses
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PuntoTendencia'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Tendencia mensual
      tags:
      - reportes
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Token de acceso con el prefijo. Ejemplo: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"